added change email adress fixed modals
