+++ app/[lang]/profile/page.tsx
@@ -2,9 +2,15 @@ import { EditProfileForm } from '@/app/components/shared/editProfileForm/compone
 import { getDictionary } from '@/app/dictionaries';
 import { authConfig } from '@/configs/auth';
 import { prisma } from '@/prisma/prisma-client';
-import { deleteProfile, editUser } from '@/services/usersActions';
+import {
+  deleteProfile,
+  editUser,
+  editUserEmail,
+} from '@/services/usersActions';
 import { User } from '@prisma/client';
 import { getServerSession } from 'next-auth';
+import { signIn, signOut } from 'next-auth/react';
+import { redirect } from 'next/navigation';
 
 type UserWithoutPassword = Omit<User, 'password'>;
 type UserWithoutEmail = Omit<User, 'id'>;
@@ -18,6 +24,10 @@ const Profile = async ({ params }: any) => {
   const { lang } = params;
   const session = await getServerSession(authConfig);
   const dict = await getDictionary(lang);
+
+  if (!session) {
+    redirect('/signin');
+  }
 
   const findUser = await prisma.user.findFirst({
     where: {
@@ -40,6 +50,14 @@ const Profile = async ({ params }: any) => {
     await deleteProfile(id);
   }
 
+  const editEmail = async (email: string) => {
+    'use server';
+    const userData = { id, email };
+
+    await editUserEmail(userData);
+    console.log(session, '>>> sesson on server');
+  };
+
   return (
     <main className="bg-bg dark:bg-bg-dark flex flex-col justify-center items-center px-4">
       <h1 className="mb-12 text-3xl">
@@ -49,6 +67,7 @@ const Profile = async ({ params }: any) => {
         className={''}
         editProfile={editProfileAction}
         deleteProfile={deleteProfileAction}
+        editEmail={editEmail}
         userData={userWithoutPassword}
         lang={lang}
       />
