+++ next.config.mjs
@@ -2,7 +2,11 @@
 const nextConfig = {
 	images: {
 		domains: ['lh3.googleusercontent.com', 'avatars.githubusercontent.com'] //Next.js разрешит использование изображений с этого домена
-	}
+	},
+	// webpack: (config) => {
+	// 	config.externals = [...config.externals, "bcrypt"];
+	// 	return config;
+	// },
 };
 
 export default nextConfig;
